requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
langgraph
langchain-core
